Minor refactoring

+++ src/main/java/com/devbliss/gpullr/controller/dto/UserSettingsDto.java
@@ -10,7 +10,7 @@ public class UserSettingsDto {
     DESC
   }
 
-  public Integer id;
+  public long id;
 
   public OrderOptionDto orderOptionDto;
